requirements.txt
fastapi
fpdf2==2.7.8
hume
hume[microphone]
protobuf==3.19.6
pydantic
pymongo==4.3.2
pytest==7.0.1
python-dotenv==1.0.1
qrcode==7.4.2
Requests==2.31.0
resend==0.8.0
streamlit
streamlit_pdf_viewer==0.0.10
uvicorn==0.29.0
google.generativeai
py3-tts

